+++ unittests/allplatformstests.py
@@ -41,7 +41,7 @@ from mesonbuild.mesonlib import (
     BuildDirLock, MachineChoice, is_windows, is_osx, is_cygwin, is_dragonflybsd,
     is_sunos, windows_proof_rmtree, python_command, version_compare, split_args, quote_arg,
     relpath, is_linux, git, search_version, do_conf_file, do_conf_str, default_prefix,
-    MesonException, EnvironmentException, OptionKey, ExecutableSerialisation, EnvironmentVariables,
+    MesonException, EnvironmentException, OptionKey,
     windows_proof_rm
 )
 from mesonbuild.programs import ExternalProgram
@@ -4678,31 +4678,3 @@ class AllPlatformTests(BasePlatformTests):
             self.assertNotEqual(olddata, newdata)
             olddata = newdata
             oldmtime = newmtime
-
-    def test_scripts_loaded_modules(self):
-        '''
-        Simulate a wrapped command, as done for custom_target() that capture
-        output. The script will print all python modules loaded and we verify
-        that it contains only an acceptable subset. Loading too many modules
-        slows down the build when many custom targets get wrapped.
-        '''
-        es = ExecutableSerialisation(python_command + ['-c', 'exit(0)'], env=EnvironmentVariables())
-        p = Path(self.builddir, 'exe.dat')
-        with p.open('wb') as f:
-            pickle.dump(es, f)
-        cmd = self.meson_command + ['--internal', 'test_loaded_modules', '--unpickle', str(p)]
-        p = subprocess.run(cmd, stdout=subprocess.PIPE)
-        all_modules = json.loads(p.stdout.splitlines()[0])
-        meson_modules = [m for m in all_modules if m.startswith('mesonbuild')]
-        expected_meson_modules = [
-            'mesonbuild',
-            'mesonbuild._pathlib',
-            'mesonbuild.utils',
-            'mesonbuild.utils.core',
-            'mesonbuild.mesonmain',
-            'mesonbuild.mlog',
-            'mesonbuild.scripts',
-            'mesonbuild.scripts.meson_exe',
-            'mesonbuild.scripts.test_loaded_modules'
-        ]
-        self.assertEqual(sorted(expected_meson_modules), sorted(meson_modules))
